Update HybridAuthAuthenticate.php
This is what I had to do to avoid "Call to a member function primaryKey()" issue when I was trying to do social auth for twitter with config settings as below. 

'Twitter' => [
                'enabled' => true,
                'keys' => [
                    'key' => 'xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx',
                    'secret' => 'xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx'
                ],
                'includeEmail' => false // Only if your app is whitelisted by Twitter Support
            ]

But application gave me "Call to a member function primaryKey()" so after doing this proposed change I was able to do Authentication Properly.